# swagger.yaml
openapi: 3.0.0
info:
  title: RS Urip Sumoharjo API
  contact: {}
  version: "1.0"
servers:
  - url: https://rs-urip-sumoharjo-api.herokuapp.com/api/v1
    variables: {}
paths:
...
  /facilities/{id}:
    get:
      tags:
        - Facility
      summary: Get single Facility
      description: Get single facility
      operationId: GetsingleFacility
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
    delete:
      tags:
        - Facility
      summary: Delete Facility
      description: Delete a facility by id. Administrator only.
      operationId: DeleteFacility
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
    patch:
      tags:
        - Facility
      summary: Update Facility
      description: Update a facility. Administrator only.
      operationId: UpdateFacility
      parameters:
        - name: id
          in: path
          description: ""
          required: true
          style: simple
          schema:
            type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
                - title
                - category
                - info
                - img
              type: object
              properties:
                title:
                  type: string
                  example: I will survive again
                category:
                  type: string
                  example: rawat-jalan
                info:
                  type: string
                  example: Melihat gelombang gelombang frekuensi yang disebut ultrasound dari bilik-bilik jantung.
                img:
                  type: string
        required: false
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
components:
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...